Improve service type safety
